Add files via upload
added a new project made as an exercise to practice what I've been learning thus far regarding Excel and Statistical analysis. In this task, I used a given dataset, and had to evaluate the correlation between Alcohol consumption and the GDP per capita of all countries of the world. As an additional quest of mine, I decided to analyse the data divided by continents too, to see if a regional correlation would exist. The results show uncorrelation for every type of metrics examined, both for Cramer's V and Pearson coefficient; the data divided by continents did not show correlation in any continent under examination, except for Oceania, which I do not take it as a significant output given the small amount of countries in such continent.